requirements.txt
fastapi
uvicorn
pandas
transformers
torch
streamlit
requests
plotly
streamlit
python-dotenv
smtplib
email